Add check for the correct text_hidName value (has to match actual file name)
Just so people do it properly since that field is the only way for Gibbed to know the file name when extracting libraries. So if someone packs a library with a wrong text_hidName field someone who unpacks that lib will not just have a wrong field but a wrong file name as well.