Fix crash when compiling libraries without a definition
This will sktip the text_hidName <=> file name check when the file to compile does not have a definition (the hex-encoded text_hidName does indeed not match the actual file name). Ideally you would convert the hex string to the actual string value but I am lazy. If you're wroking with raw hex it's your job to not screw it up.